message-collector/requirements.txt
nltk==3.8.1
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
python-telegram-bot==21.1.1
